chore: use tsx, node --test
