+++ .turbo/config.json
@@ -1,5 +1,4 @@
 {
   "teamid": "team_ducktors",
-  "apiurl": "https://eu.trrc.ducktors.dev",
   "token": "turbotoken-oss-europe1"
 }
