+++ packages/abstract/src/index.ts
@@ -1,2 +1,2 @@
-export { Storage } from './abstract-storage'
-export type { AbstractStorageOptions } from './abstract-storage'
+export { Storage } from './abstract-storage.js'
+export type { StorageOptions as AbstractStorageOptions } from './abstract-storage.js'
