+++ packages/gcs/src/index.ts
@@ -1,2 +1,2 @@
-export { Storage } from './gcs'
-export type { StorageOptions as GoogleCloudStorageOptions } from './gcs'
+export { Storage } from './gcs.js'
+export type { StorageOptions } from './gcs.js'
