+++ packages/local/src/index.ts
@@ -1,1 +1,2 @@
-export { Storage } from './local'
+export { Storage } from './local.js'
+export type { StorageOptions } from './local.js'
