--- packages/s3/tsconfig.build.json
@@ -1,9 +1,0 @@
-{
-  "extends": "@ducktors/tsconfig",
-  "compilerOptions": {
-    "outDir": "dist",
-    "sourceMap": true
-  },
-  "include": ["src/**/*.ts"],
-  "exclude": ["src/**/*.test.ts"]
-}
